backup: Increase main storage size from 100 to 120 TB
Bacula has ran out of size for its main storage location, making
backups getting stuck. Increase the maximum amount of volumes
(after resizing online its available filesystem) from 200 to 240,
which for half a TB volumes, that means a 20 TB size increase.

To prevent this in the future, we should speed up the split of
gerrit and gitlab backups to a separate host, like we did with
databases.

Change-Id: I7478da8ec5e1c07b16577d7a722cb7d9532c6521