sretest2010: set to be installed like a new ms-be* node
We're taking this host over for data-persistence testing; it's a
single-CPU system like our existing Config-J ones.

This change should make all the necessary changes to get this node set
up as a swift backend server with JBOD disks. The aim is to do
everything in one CR so it can be straightforwardly reverted once
we're done testing.

Bug: T394357
Change-Id: Id4ad51edf1f2a2862a5b1e5316a61b18fe995b4b