pin cloudcephmon1004 to version 'pacific'
Since we still have osds running p, we don't want a mon
with n+2. I will leave this host on 'pacific' for now,
which will be n-1, then it can skip to reef after the osds
are on quincy.

(Context: bullseye refuses to install properly on this host, and quincy isn't available on bookworm)

Bug: T402190
Change-Id: I15d2af351c6fd10d037d71c85caaf94f1ac467b1