trafficserver: relax service lookup in mw-next-routing
As part of T402412, we can no longer assume that only services with
-next equivalents will have been selected by scripts earlier in the
plugin chain.

While mw-next-routing already handled this correctly (i.e., early exit
with NO_REMAP), it did emit an error, since prior to now this would have
indicated a misconfiguration.

Instead, do not log an error when this occurs.

Bug: T403655
Bug: T402412
Change-Id: I6053db65b2b72ef0d81281bd0ea2ec16d481619d