zuul(new): update path to ssh private key for zuul executor
for the new zuul-executor-setup this ssh key is what the zuul
executor uses to access the nodepool node.

quote: "The running container will need a bind mount
with an ssh key allowing the zuul user to access the nodepool node"

quote: "Bind mount a directory with the global nodepool ssh private key
(if you want, you can put this under /etc/zuul or /var/lib/zuul so you don't
need an extra mount point)"

Based on that we put this into /etc/zuul/ssh/id_rsa in T403847#11153108.

The path /var/ssh/nodepool does not exist.

Bug: T403847
Change-Id: I377112f8c30ff8f2a98a6feb932310b712eb6051