docker: add support for trixie, ensure docker-cli is installed
On Debian trixie, /usr/bin/docker was moved from the docker.io
package to a separate docker-cli package.

docker.io recommends docker-cli, so by default would be pulled in
but since we set Install-Recommends "false" globally that does not
happen for us.

The result is that puppet runs and docker.io installs without issues
but then the user is surprised with a "docker: command not found".

Therefore, check if the package name is set to docker.io and we are
on trixie, and if so explicitly install the docker-cli package as well.

Unless we want to question that we disable recommends. Some Debian
support channel say it would be a bad idea because it creates traps
like this.

Change-Id: I6a0d3ff19dfc509606290d57841f30588d54134e