ssh: Disable X11 for the new-style sshd.d template
Disable X11 forwarding for the new-style OpenSSH configuration template.

This is disabled in OpenSSH upstream by default, but enabled in the
Debian default configuration.

There's not much of a security impact, but it's functionality we don't
use or need, so it makes sense to disable it regardless.

Bug: T400478
Hosts: sretest1002.eqiad.wmnet, sretest1003.eqiad.wmnet
Change-Id: I23abdacf00874c88055012a1035e545ecadb8341