Switch the wdqs-internal services from http to https
We have started using TLS for the wdqs services, but this has never been
applied to the wdqs-internal-main and wdqs-internal-scholarly services.

They all have the required envoyproxy services up and running, so all we
need to do is change the target port for the services in the service
catalog.

Hosts: P:spicerack,P:lvs,P:liberica
Bug: T193473
Change-Id: I748a97380bd133a5d9e2a1542182d3cf8a169de9