varnish: Switch Commons/Googlebot pilot from desktop to unified mobile
Follows-up If586d4e453eee05576dc, which disabled the mobile redirect
for Googlebot on Commons as an early experiment for unified mobile
routing. That patch disabled the redirect but also switched the
response HTML from mobile to desktop, to a "Googlebot Mobile"
crawler.

To minimise the period during which Google is indexing Commons based
on desktop HTML, switch it back to mobile HTML now that the unified
mobile routing is available.

Don't rollout unified mobile routing to Wikimedia Commons in general
yet, that will happen later in T403510.

Bug: T397267
Change-Id: Ibaac4e7486616b428eee2f7d5375133742a33209