Revert to using a hostname for the druid_poublic coordinator
We had attempted to switch away from a hostname to a load-balanced
endpoint for the druid_public data purge jobs.

However, this did not work because we do not yet have the
druid-coordinator service under LVS. It is only the druid-broker service
that is load-balanced.

For now, we will revert to using a hostname, but we will look at whether
we can use LVS for the coordinator service, as well.

Hosts: O:analytics_cluster::launcher
Bug: T403955
Change-Id: I89dd3a387d5978a33191f6528f507e0225deac5c