deployment_server: Add a script for mass-deploying helmfile services
charlie.py recursively searches out helmfile services in a given
repository or subtree, identifies all the environments for each service,
and diffs or applies them by shelling out to helmfile. It's named after
Charles Babbage because it's a labor-saving mechanized difference
engine.

(This is associated with the current round of Envoy upgrades because
it's a necessary component of deploying the image and configuration
changes to all services.)

Bug: T380211
Change-Id: I6a6a6964153084632f33b1515fae2ae9c6753a31