deployment_server: allow different namespaces to be deployed within a group
The existing code assumes that kubernetes clusters within the same
cluster group (ex: eqiad/codfw both belong to the 'main' cluster group)
will all have the same namespaces.

That hypothesis does not hold in the dse-k8s cluster group, where
dse-k8s-eqiad has much more workers, and workloads that we don't expect
to run in a multi-dc fashion (airflow, etc).

We tweak the service config generation as well as the service
certificate generation, to perform an overlay of the cluster_group and
cluster-specific services.

For example, if `opensearch-operator` is defined under `dse-k8s`, then
the service will be deployed in both dse-k8s-eqiad and dse-k8s-codfw.
However, if airflow-test-k8s is defined under `dse-k8s-eqiad`, then it
will not be deployed to dse-k8s-codfw.

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Bug: T404068
Hosts: O:deployment_server::kubernetes
Change-Id: Ib4a1b2f9e3c442a164653dd8949f36a8b1143def