zuul::executor: drop bootstrap file, adjust systemd service command line
We don't need a bootstrap script after all; if we just run docker
run with the registry URL and a specific image version.

This should do the right thing and pull if needed but only run all
following runs.

Avoiding an "init" docker pull to setup a new executor instance.

Also make the ExecStart line more readable and link to our ticket.

And finally, small but important, fix a typo in the version string!

Bug: T403847
Change-Id: I5eacb1b0b9f44def4e0ac5e5efaf2f8463480d37