zuul::executor: configure exit code 143 as success in systemd unit
A container that is stopped by systemctl and exists with code 143 has
been SIGTERMed properly and without an issue.

But systemd treats any non-zero exit code as a failure, which would
turn into a failed systemd unit and degraded systemd state.

Ultimately this would mean we get a Phabricator ticket from monitoring
each time we stop the service. :p

Good that we can configure systemd to tell it which exit codes are
not failures.

ref: https://www.groundcover.com/kubernetes-troubleshooting/exit-code-143

Bug: T403847
Change-Id: I6b6669e99e74558fb063997820c63871f3d53aab