Cephosd: revert to manually setting up peering IPs
This reverts commit 7109c23e7b697e6d77a11b85b12edfa616b23220.

Reason for revert: Unfortunately as things stand this config causes
BFD to fail on IPv6 when the peers are directly adjacent.  Bird
assumes multi-hop as it is sourcing from one network (global unicast
IP) connecting to another (link local on router).  Network automation
is forcing 'single-hop' mode in the router configuation and things
fail.

This doesn't happen on other anycast roles which peer from their
global unicast to the CR loopbacks - so automatically multi-hop.  NS2
in esams is v4-only.

It _should_ be fairly easy to fix this but right now I don't have time
to look, and it results in BFD down alters.  So let's keep it on
global unicast <-> global unicast, which is sensible and just works.

Change-Id: I3c896b099e2e7ee82cb9633180befb87ea3face8