trafficserver: multi-dc: use client request host rather than remap
In the current logic used by multi-dc.lua, the remap rule will always
be obeyed for writing the base or -ro hostname. This is not the
desired behaviour, as we wish to have multi-dc as part of a chain
where gateway-check.lua is called earlier in order to allow us to
rewrite the hostname on a URL match, and then use the -ro version of
that host (in this case, rest-gateway-ro). This is a pattern we use in
mw-next-routing.lua.

This change uses get_url_host in order to rewrite the
already-rewritten hostname from earlier changes in the chain. There is
a risk in this change if this rewrite pattern does not happen as
expected when we call set_url_host. We also need to ensure that the
top-level `replacement` call is in place before we start calling this
chain of rewrites, which it seems fairly certain is happening in order
for us to use the same set/get behaviours in mw-next-routing.

Bug: T401396
Change-Id: Iec8b3f2e98fee028c96f99910ed4eeb71b12e3e7