profile::amd_gpu: add support for amd-smi from ROCm 6.4.3
In various tests for the new MI300X GPUs I found out that only
the ROCm 6.4.3 suite seems to be able to properly support features
like GPU partitioning.

Add a repo config for ROCm 6.4.3, limiting the packages to what
really necessary, and add an option in the amd_gpu profile to
deploy them.

Bug: T403697
Change-Id: Id5420b28cc47e07f412002d7c085c2c041ab49fd