installserver: Select the custom nginx provider with no additional modules
The install servers are now on bookworm so we can start using the new nginx
configuration scheme.

The installserver site doesn't use any extra modules, so we can just go
with the empty module list.

(Strictly speaking the empty module is already the default via
common/profile/nginx.yaml, but it still seems useful to have it
here more expressively along with the configuration setting to select
the variant).

Hosts: install1005.wikimedia.org
Bug: T329529
Change-Id: I1e0cbf2a5bde3fa8d3f7640dc41f34753f6a9ca5