Add kubernetes tokens for an analytics-test namespace on dse-k8s
This is a new set of tokens for carrying out work in the analytics-test
namespace around spark applications.

They will be created for both dse-k8s-eqiad and dse-k8s-codfw since both
will be useful targets for testing, even though Hadoop etc are only
available in eqiad.

The tokens will be deployed with the privileges configured so that only
members of analytics-deployers will be able to read them.

Hosts: O:deployment_server::kubernetes
Bug: T381416
Change-Id: Ic756341ddb6d6ccb434d9014a27bbda40df9b819