deployment_server: Add --resume_at to charlie
Charlie acts on all services alphabetically, it's a long list, and
sometimes a run needs to be interrupted partway through the alphabet.
When some diffs were skipped the first time around, it's annoying to
have to skip them the second time. Instead we can use --resume_at to
pick up where we left off.

Change-Id: I625bdfaa276f82abf8ede142aaac2a706a6a6964