deployment_server: Add retry options to charlie
When diffing or applying returns an error, offer to retry. Sometimes
this helps when an apply times out due to a transient health check
issue, other times there's a real problem but the user can fix it in
another terminal and then continue.

Consequently move some error handling up from diff_and_apply() to
multiapply(), which smooths out some of the try-excepts.

Change-Id: I0f1540153d03e1c63e76b6a9a32bb3936a6a6964