varnish: Remove X-Cache-Status -remote values
In the before times, there was "front" (Varnish frontend), "local"
(Local varnish backend cache), and "remote" (Core DC Varnish backend
cache). "Remote" meant that the backend cache was fetched from another
DC.

This is no longer the way things work!

As a result of all the changes in infrastructure, we only ever receive
-local and -front. As a result, -remote has been made redundant.

The comment enumerating all the possible values has been fully removed
instead of updated because there exists the possibility of "int-tls",
which is set by HAProxy. As this is lies outside of the VCL itself, we
should just remove the implied expectation that only Varnish will be
setting that value.

Bug: T404654
Change-Id: I027a1581381df263e28a2a310b3b50f5dbd329be