Revoke analytics-wmde-users membership from conniecc1
Following a review by WMDE staff, it was determined that conniecc1 no
longer requires access to the analytics-wmde-users group.

Membership was correctly assigned at the time, but is now superfluous.

The user has been consulted and is in agreement. There is a Slack thread
here: https://wikimedia.slack.com/archives/C055QGPTC69/p1758616025287669
that relates to this request.

Bug: T228447
Change-Id: Id5d7e5653f242cfa2726d196617c45f9bd1b0c1f