cirrussearch: re-enable CPU performance governor
Today's datacenter migration resulted in some added latency for the
CirrusSearch cluster (see also T405394). During the troubleshooting, I
noticed at least one of the hosts was CPU-bound , so I went to check on
the CPU performance governor for our CODFW cirrussearch hosts.

Unfortunately, this revealed that none of our CODFW cirrussearch hosts
( cirrussearch[2061-2115].codfw.wmnet ) have the performance governor
enabled. In ticket T386860, we received approval from DC Ops to enable
the governor on a subset of hosts , but this setting was lost when we
renamed the elastic hosts to cirrussearch.

We've received approval from DC Ops to re-enable the performance governor
on these hosts, so let's add the Puppet code we need.

Hosts: cirrussearch2083.codfw.wmnet
Bug: T386860
Bug: T405396
Change-Id: I27ee87f42d2cd32a5ed308e83fe65900aa4806f6