admin: upgrade ericmill from ldap_only to a-privatedata with no shell
Per T404903#11211475 ff this existing "LDAP only" user needs to be
added to the analytics-privatedata-users "shell" group but WITHOUT
giving them actual shell access.

So no SSH key is added but the user is moved to the so-called shell
user section without having actual shell access.

This is the next access level above just membership in the "wmf"
LDAP group per:

https://wikitech.wikimedia.org/wiki/Data_Platform/Data_access#Access_Levels

Bug: T404903
Change-Id: Ie942cec23d3c9acfea69231e8670759ab73ade00