imposm-initial-import: Set service passwords
Starting with Postgres 15 the hashing changed from md5 (for which we could
precompute the hash in the postgresql user defines) to scram-sha1, which
gets salted with a value we've so far been unable to extract to compare
the hash in Puppet, so password changes are currently not supported until
T326325 is resolved.

But effectively we only need to set them once, so do it as part of the
initial OSM import.

Bug: T381565
Change-Id: Iec4f7a668db714c4a04e8b4c760c0af29ea79c92