dbbackups: Partial revert of dbprov1007 setup, back to dbprov1003
dbprov1007 disk setup was incorrect. This was only caught when
backups started failing due to lack of disk space (the first ones
worked ok).

This was not due to a software or file configuration issue, but
due to a mistake on disk installation.

Reverting generation of eqiad backups so they return to be done
from dbprov1003. dbprov1007 backups worked correctly.

Silencing and removing monitoring from dbprov1007, which will need
a reimage to be setup as intended.

This is a partial revert of:
a6e00a7b7610d63c0ae9ca817766b38807777a69

Bug: T403166
Change-Id: If288b7dff5aeb165bcb7cf330cc37ba09539a929