phabricator: hiera'ize the apc_shm_size variable
To be able to set it to a different value for the test instance.

We can't hardcode the values that work in production when the test
instance has much smaller RAM.

It would break and that's why we currently have puppet disabled there.

This should allow us to re-enable it. Plus for any future changes of
the value in production we don't have to touch puppet code anymore.

Bug: T401157
Change-Id: Ib4dce58bfaedab56df25a00aecc7c35ec8fbb03a