dbbackups: Upgrade db1245 to MariaDB 10.11 so it can take over db1150
db1150 crashed due to a memory error. Move backups generation to
other hosts while the db host is serviced and silence notifications.

s3 backups for now will be generated from db1240
s4 backups for now will be generated from db1245

In order for db1245 to take backups, it will have to be upgraded to
MariaDB 10.11.

Bug: T405885
Change-Id: I7ac58035c8f06f0823c430e1af8202b932fdde4c