tools: add more reliable stats on nfs stuck workers
This uses the fact that lsof gets stuck on nfs to signal the worker
as stuck and start recording the processes that get into the
uninterruptible sleep (D) state.

This should give us a better metric than just checking if there's any
processes in D state, allowing to clearly see if there's any worker
stuck.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: If6adaea3f26a64f5cd7215ec56a88430ef6e5757