aptrepo: add kubernetes 1.23 support to Trixie Wikimedia
The ML team needs to test their new GPU nodes as Kubernetes workers
on Debian Trixie. They are not ready to migrate to Kubernetes 1.31
yet, so the K8s SIG decided to allow 1.23 to be deployed on Trixie
(workers only).

The plan is to create the necessary components and copy the binaries
from bookworm-wikimedia.

Bug: T405891
Change-Id: If447bcb86a7b933f1f64bad7c50a9f589b85eebd