zuul: move new zuul nodepool setup to its own profile
The "zuul::main" profile is become a bit large and before
we are going to add more code to it for T405118 and T405119,
break out the nodepool part into its own profile.

Then include that profile in the zuul::main role and adjust the
Hiera keys.

That's how it should be. A role should be composed of separate
profiles and having one profile match each of the tasks we
created also makes it more readable.

Bug: T395938
Change-Id: I745f8c87b4c57f4d64bf44ab8fbb5290f19b1530