role::maps::master: enable planet sync in codfw
We turned off the planet sync in codfw since we have been working
on other (newer) stacks, hoping to do a quick upgrade.
It turned out to be a very long journey, and we have been running only
in one datacenter since then.

Turn on planet sync replication to allow maps in codfw to catch up,
and to be used again in production when needed.

Bug: T381565
Change-Id: Ie0529358b377697de0141e5244ad40abc94fc07f