deployment_server: Prefix `helmfile apply` output with "[service env]"
Previously we did this the easy way with `helmfile diff`, by using
subprocess.run. That method blocks until the command completes, so we
were buffering all the output first and then printing it out. With
`apply`, we don't want that; it takes longer to run and the user needs
to monitor its progress. But the prefix is still useful for keeping
all the output straight, especially in the long scrollback.

So, switch from using run() to using Popen() directly (for both use
cases), which lets us stream from its stdout and stderr a character at a
time.

As a result, _exec_helmfile() no longer returns a CompletedProcess;
instead it returns a tuple of the exit status and output, which is all
we were using. The output is already prefixed with the service and
environment, which might be unhelpful if we ever start to inspect the
diffs' contents or compare diffs *between* environments, but we can
address that if we get there. In the meantime it simplifies existing
usage.

Change-Id: I803378202e8e52dca50fa210285da4256a6a6964