Customise the login.html template of JupyterHub to hide the TLS warning
Since our JupyterHub users are already using an SSH tunnel to access
their servers, the warning about recommending the use of TLS is moot.

Therefore, we can extend the default template for the login.html page
and configure it not to display this warning, by the use of jinja
template overrides.

Hosts: O:statistics::explorer,O:analytics_test_cluster::client
Bug: T403863
Change-Id: I76c080f3a9051d08d2072f83c6529d246d30c910