varnish: Enable unified mobile routing on idwiki, frwiki, dewiki
Update "old-style" tests to target enwiki, because dewiki is now
a "new-style" wiki (already covered by tests, using testwiki as host).

Split out mobile/desktop responses from mock server.
These two should (and do) share a cache entry on old-style domains
for non-redirect URLs (such as /w/load.php), and this was the case
in the test as well for de.wikipedia.org. It is not the case for
en.wikipedia.org in the test, because an xLab mock is entering
the 08-mobile-hostnames-rewrite.vtc test, which has an experiment
active on en.wikipedia, and happens to assign these into requests
different buckets, and thus not share a cache.

Before
> ReqHeader  User-Agent: Android
> …
> ReqHeader  Host: de.wikipedia.org
> ReqURL     /w/load.php?modules=startup
> ReqHeader  X-Experiment-Enrollments:
> …
> VCL_return hash
> VCL_call   HASH
> VCL_return lookup
> VCL_call   MISS
> …
> ReqHeader  User-Agent: Firefox
> …
> ReqHeader  Host: de.wikipedia.org
> ReqURL     /w/load.php?modules=startup
> ReqHeader  X-Experiment-Enrollments:
> …
> VCL_return hash
> VCL_call   HASH
> VCL_return lookup
> Hit        1026 287.771683 1200.000000 0.000000
> VCL_call   HIT

After
> …
> ReqHeader  Host: en.wikipedia.org
> ReqURL     /w/load.php?modules=startup
> …
> ReqHeader  X-Experiment-Enrollments: 2025-abtests2-foo-10x10=grp0;
>            2025-abtests2-bar-5x10=grp2;2025-abtests2-sel1-4x25=grp2;
>            2025-abtests2-sel2-4x25=grp2;
> …

Bug: T403510
Change-Id: I712352fc235bc9e6db541fa29296d85ee30ecc39