deployment_server: set the php.version value in mwscript-k8s
Since mw-script helmfile releases do not use the scap-managed values
files, they currently fall back to the chart default for php.version,
which will be inaccurate while a migration is in progress.

Instead, set it in the values file synthesized by the wrapper, based on
the --php_version option used to select an image.

While it won't be possible to infer a correct PHP version in the event
the user has set --mediawiki_image, that's sufficiently rare that it's
likely not worth worrying about. In this case, we set nothing.

Change-Id: I5479d5faf698bcdd5429194c7a6d691597665790