Add an-launcher1003 to the list of permitted rsync and nfs clients
We are adding a new analytics launcher, called an-launcher1003, to
replace an-launcher1002.

This commit enables the new host to exchange files with stat servers
using rsync, as an-launcher1002 is able.

It also configures the clouddumps100[1-2] servers to permit
an-launcher1003 to mount the NFS shares.

Hosts: O:dumps::distribution::server
Hosts: O:statistics::explorer
Hosts: an-launcher1002.eqiad.wmnet,an-launcher1003.eqiad.wmnet
Bug: T402943
Change-Id: I64ebb9f0445338567485856157eee0603e216e9e