deployment_server: ensure all users can traverse service private directories
This does not mean that they are able to list or read any of the
contained files, but only traverse the directories. This is useful in
dse-k8s-eqiad as airflow private files are group-owned by
airflow-deployers, which falls under the `o/other` permissions bits of
the parent directory, currently set to 0.

I'm suggesting taking this direct approach given that we already
went through https://gerrit.wikimedia.org/r/c/operations/puppet/+/1151657

Signed-off-by: Balthazar Rouberol <brouberol@wikimedia.org>
Hosts: O:deployment_server::kubernetes
Change-Id: If7f74aea0c6d97675372c58786f053f5c5bc0346