gitlab: disable and remove partial backup
partial backups which excluded packages and artifacts were
introduced as a workaround for long-running backups.
They helped keep the replicas in sync without blocking
the production host for multiple hours.

With T378922 packages and artifacts are no longer stored on the
host. So the regular backup is much faster and replaces partial backups.
So this change removes partial backups and increases the schedule
for the full backup

Bug: T378922
Change-Id: I54d0ae7f5dae3f433ffa17e096bcb9e1e22d38b8