jenkins: move apt-get command for scap to a dedicated file
On some jenkins hosts (releases) scap will run apt-get update
to update jenkins.

There is a sudo privilege line to allow that but there were problems
with it and escaping.

To fix that issue and as general best practice, move the command
to a file in /usr/local/bin and adjust sudo privileges to run that
file.

We tested on releases2003 and it fixes the upgrade; it installs
jenkins also when no jenkins is installed at all yet.

Bug: T405352
Change-Id: I6e1b974ffa8d33c1c606a08969f5f789518a63e4