zuul: use mysql+pymysql instead of mariadb+mariadbconnector in db URI
Use the SQLAlchemy dialect mysql with the driver pymysql instead
of the mariadb dialect with the mariadbconnector driver when connecting
to the new zuul database.

While the database we are using is certainly mariadb, SQLAlchemy also
says:

"The MariaDB variant of MySQL retains fundamental compatibility with
MySQL’s protocols however the development of these two products continues to diverge.
Within the realm of SQLAlchemy, the two databases have a small number of syntactical
and behavioral differences that SQLAlchemy accommodates automatically.

To connect to a MariaDB database, no changes to the database URL are required:"

and we don't have the mariadbconnector driver in the zuul-scheduler
image.

Starting it fails with "No module named 'mariadb'" but using mysql+pymysql
which is also supported by zuul appears to work normal.

Bug: T405118
Change-Id: I73f5588aab4d56a3fc0ea3cadfe92d1e1257ceb6