Add the JupyterHub.template_paths value to the config file
We have previously added a configuration statement to puppet that was
intended to configure JupyterHub to use our custom template file.

However, this was not automatically rendered into the configuration file
as expected. This commit adds a specific configuration parameter to the
jupyterhub_config.py template, so that the value will be enabled.

Hosts: O:statistics::explorer, an-test-client1002.eqiad.wmnet
Bug: T403863
Change-Id: I5383645daadf4453b46ce6d22ec8c53b50deee80