Remove the airflow profile from the analytics_cluster::launcher role
We used to run the airflow-analytics instance on an-launcher1002 but
that has now been superseded by Airflow on Kubernetes.

We are also deploying a new analytics_cluster::launcher host called
an-launcher1003. This host runs bookworm, but as it cannot find the
airflow package in the bookworm repository, installing it with puppet
fails.

We can therefore remove the profile from this role.

Hosts: an-launcher1002.eqiad.wmnet,an-launcher1003.eqiad.wmnet
Bug: T402943
Change-Id: I9f81f7ec00a2d7b65523f99068ff01c781cc0b2e