P:cache::haproxy: start preparing for known-client DSL
This patch starts preparing the tls_terminator.cfg.erb template-template
for the introduction of the known-client identification DSL fragment.

First, this sets the stage for x-provenance and x-trusted-request data
flow in / out of the fragment [0]. The latter consumes req.provenance
and may update both req.provenance and req.trusted_request.

In order to support updates to the former, x-provenance is now set later
on, after where the known-client DSL will be inserted (no such change is
needed for x-trusted-request).

Second, it prepares the x-trusted-request "C" scoring logic for the
existence of "B" scored requests, which should not be downgraded.

Third, it shifts ja3n/4h computation and reporting earlier, just before
where known-client DSL will be inserted. Together with some follow-on
changes to shift x_requestctl logging setup earlier, this ensures that
they will be logged, even when requests are denied due to impersonation
of a known-client.

[0] For an simple example of what the DSL fragment will look like, see
    the description on:
    https://gitlab.wikimedia.org/repos/sre/hiddenparma/-/merge_requests/120

Bug: T403220
Change-Id: I4122c3a31084a3cd31f801095565786975bb15e9