wikistats: do not ensure dir that is already used with git::clone
The git::clone defined type already has an internal ensure
for the directory.

There is an existing FIXME comment that git clone itself should
do it but it's also there to ensure owner/group do not change.

(line 196ff modules/git/manifests/clone.pp)

Either way it needs to happen AFTER a git::clone or git cloning
fails because the directory already exists.

And doing it a second time outside of git::clone is definitely not
needed then and just causes the same issue on an initial puppet run
when setting up a new instance.

Bug: T401859
Change-Id: I1da6200e1200cc849a1740335e2b30231b18fb4f