P:cache::haproxy: move x_requestctl setup into listen section
Previously, this logic was replicated in each requestctl backend section
definition.

Since we will soon be injecting a fragment of requestctl-generated DSL
into the listen section, which will similarly be logging outcomes to the
txn.x_requestctl var *and* may deny requests (impersonation), it makes
sense to deduplicate and lift this into the listen section.

While doing so, I've also added a bit more detail to the associated
comments to capture what's happening.

What are the implications of this change?

Recall, http-after-response rules are evaluated in declaration order in
the section they appear. There are two important caveats relevant to our
use case regarding rules declared in a backend section:
1. They are applied before rules in the listen section.
2. They are never applied if processing never switched to the backend
   via a use_backend (e.g., request was denied earlier than that).

So, in the happy case where we *did* switch to a requestctl backend,
this is equivalent to the prior behavior - i.e., it just shifts updates
to x_analytics "earlier in declaration, later in execution" in a way
that should not have an effect on its value.

Where this gets interesting is the case where we don't get x-analytics
back from Varnish (so x_analytics is not set by the http-response rule
that initializes it from the header) *and* we never switch to a backend
that would have otherwise synthesized one - i.e., for trusted requests
or silent-drop.

For the trusted request case, we'll now get a synthetic x_analytics in
logs (and indeed one that has a ja3n/4h, since we synthesize *before*
their append-only updates).

For silent-drop, there is no difference, since http-after-response rules
aren't executed at all for silent-drop (or reject), so x_analytics does
not get populated either way.

Hosts: O:cache::text
Hosts: O:cache::upload
Bug: T403220
Change-Id: Id83196cc5a72b57dce3e902d3d0959381db17e01