cpufrequtils: improve cpupower's config
The current systemd unit doesn't work, since the cpupower
binary expects a certain amount of parameters to run.
In https://bugs-devel.debian.org/cgi-bin/bugreport.cgi?bug=894906
the patch proposes a handy bash script to add under /usr/libexec,
that can set various tunables based on the environment variables
provided in the running env.

It seems a good way to proceed, to remain as close as possible
with the last proposal from the upstream community.

Bug: T405891
Change-Id: Ifdf687098f85ed8816831333f68052d087ceb4a2