backups: Migrate Gerrit and GitLab backups to new storage hosts
The production jobs are getting overloaded and filled up with
new data. Split the jobs into the general storage jobs, keeping
the name "production", and the new, dedicated for code and
related repositories data (for Gerrit and Gitlab) storage "repos".

This is the final part of the migration, when we apply the new &
tested configuration to both services.

Bug: T403946
Change-Id: Ifdc78825b7834129d8238e9c1ec44eac27086b9c