Add the analytics-research keytab to the stat boxes
In order for members of the analytics-research-admins group to be able to
interact with kerberised services as the analytics-research system user, we
need the analytics-research keytab to be present on the stat servers.

This is in order to support commands such as `yarn logs` for
troubleshooting purposes.

Hosts: O:statistics::explorer
Bug: T403207
Change-Id: I183372b1022f21f69f1f166609510ed1dc461044