P:toolforge::k8s::haproxy: Prefer IPv4 for backend nodes
While some Kubernetes workers are in the dual-stack network (and have
AAAAA records), the clusters themself are still IPv4-only. (This is
T380060 and is very non-trivial to solve.) And because the cluster is
v4-only, NodePort services are only offered on IPv4. So we need to tell
HAProxy to reach those services via v4 instead of trying and failing to
do so on v6.

This patch does that by configuring the "resolve-prefer ipv4" HAProxy
option. This involves patching the HAProxy config to do the DNS lookups
internally, instead of relying on the system resolver for that. The few
lines of HAProxy boilerplate to define a resolver is placed in a
separate profile to make that re-usable.

Bug: T405078
Change-Id: I6a6a6964e301ffe50232ac52ee9f03cf682eb668