varnish: Refactor 08-mobile vtc to pair req/resp assertions
This restores an assertion for Google Translate on new-style wikis
(es.wikipedia.org), to receive the mobile version.
This was lost in Ic9b1ca9f42a0cf4a when switching from 302 + location
(mobile redirect) to merely 200, which does not mean it is a mobile
version, that's asserted by the X-Subdomain header.

Similarly add mobile version checks (X-Subdomain) for wikimedia.org
wikis after removing the 302 assertion in Id162eed3af5.

Change-Id: I239d82693f80f4928e54e82630a6c7243bc90fcd