P:toolforge::proxy: Disable connection failure tracking
Having both A and AAAA records on the VIP made Nginx think there are
multiple backends, and so now it's tracking failures for each of them
and stopping traffic to a specific "backend" when it sees a single
request to it fail. When that happens for both address families in a
short enough time, we start serving 5xx errors.

This is not what we want in our setup, so this patch disables that
functionality:
- Setting parameters for the backend server requires moving it to a
  upstream{} block, so that's done.
- Configuring a resolver in an upstream block is paywalled until Nginx
  1.27 (we run 1.22 which shipped with Bookworm), so that is removed
  entirely as the server block no longer has any use for that. A
  resolver was mostly relevant with the dynamic routing setup used with
  the Grid Engine, and I believe it wasn't even used with the hardcoded
  proxy_pass statements (and instead Nginx used the system resolver to
  look those names up on startup).

Bug: T405078
Change-Id: I6a6a6964ae3a481fc263bc0203ac8c60c6d91974