Add the analytics-research-admin group to stat admins
In order for members of the analytics-research-admins group to be able to
interact with kerberised services as the analytics-research system user, we
need the analytics-research-admins group to be part of the stat hosts admins.

This is in order to support commands such as `yarn logs` for
troubleshooting purposes.

Hosts: O:statistics::explorer
Bug: T403207
Change-Id: I023624966fc66cbb8d862672dc403cb0bc8d7ea3