zuul: use zuul_main_nodes to determine zookeeper server
We need to get the IP of the zookeeper server(s) in both
main and executor zuul nodes.

So since we would lookup the same thing from multiple profiles
the value should be in a common place and not be repeated in multiple
places.

Also we already have a list of "zuul_main_nodes" in common Hiera
and we know main nodes are also zookeeper servers.

So define the first zookeeper server as the first zuul main node
and do not have lists of host names in separate places.

Finally, add the port number.

Bug: T395938
Bug: T406366
Change-Id: I9a2127c1a0bc6ca84af8dd9db20209bc9e59471b