osm: refactor swift scripts and make event-template dynamic
The bootstrap-tile-storage.sh and send-tile-expiration-events.sh
scripts dump the list of tiles from swift and pipe it for
further filtering/processing. From recent tests this led to a huge
amount of memory used on the Postgres master, ending up
in the OOM acting to clear resources.

Refactor the scripts to just dump to a temporary file. Add also
a systemd timer to clean up old files that are not needed anymore.

Make also event-template.json dynamic, since the event stream's stream
to use for tiles expiration varies across clusters.

Bug: T381565
Change-Id: Ie8c81acf6334c2dd640ab50b5df1ecc900006b0a