deployment_server: Refactor charlie to add a Service dataclass
Also: in service_inventory(), omit any service with no environments,
instead of including it with an empty environments list. (This comes up
when we have an environment or glob set with -e.) It doesn't make a
difference for any of our current use cases, except "charlie list",
since they just loop over the empty list and do nothing, but it matters
for the implementation of T406212.

Change-Id: Id7c1ffd03843b8898845c9eb7e19b5986a6a6964