zuul: create class and systemd unit for new zuul-web service
zuul-web serves the web dashboard to end users and
communicates with database, zookeeper.

We have already setup CDN and and Apache reverse proxy to be used in
front of it as part of the general "puppet zuul" task.

And since I700ef191d45dd109aab2 the zuul base class installs
the same config on main nodes and executor nodes.

Copies the same pattern we have used for zuul-scheduler/zuul-executor
but with the adjust docker image.

But only adds this service on the main nodes.

Bug: T395938
Bug: T405119
Change-Id: Idf39a636ac89e37e5ba8a034fbbd3a40bef12636