deployment_server: Add --priority to charlie
Act on the given services, in the given order, before any others. This
is important for the wiped-cluster use case, since we'll want to deploy
services with DaemonSets first, then any services with critical downtime
constraints.

When implementing --resume_at, that means we find the given service in
the actual iteration order instead of comparing them alphabetically.

Bug: T406212
Change-Id: I1dd40eaf799a30325c33f0eda263ecb86a6a6964