P:toolforge: Migrate error page static assets to tools-static
The Toolforge front proxy nginx currently serving these files is
planned to be retired entirely. It is being replaced by a HAProxy
instance, which can't serve files from the filesystem.

This patch migrates the static assets used on proxy-generated error
pages to be served from tools-static instead, to remove that role from
the front Nginx instance.

Bug: T283948
Change-Id: I6a6a696428d92d322f37dbef563a41694f418f85