profile::amd_gpu: add initial support for the k8s node labeller
The k8s node labeller is described here:
https://github.com/ROCm/k8s-device-plugin/tree/master/cmd/k8s-node-labeller

It is a kubernetes controller that is able to read from /dev,
recognize gpus and report their metadata (VRAM, type, etc..) to k8s
so every node has its set of labels. At the moment we can only
assign a GPU to a pod without targeting which one, meanwhile after
the node labeller adds metadata we'll be able to be more specific
in deployment specs.

The suggested way of running the controller is as a pod, with
permissions to access /dev and to watch/get/list/update the "nodes"
resource. Since we already run the gpu device plugin as daemon
on the bare metal workers, it seems convenient to add this binary
as well to the related Debian package and run the controller outside
k8s. This patch adds the TLS certificate material to support
the new username, and then a subsequent deployment-charts patch will
add the related clusterrole+binding.

Apply the new config to ML staging.

Bug: T373806
Change-Id: If2b6a2f78ce12aec7866d921bc5dec3e99c13f27